Create locust tests factory
